requirements.txt
google-assistant-grpc==0.2.1
google-auth-oauthlib>=0.1.0,<0.3
urllib3[secure]>=1.21,<2
sounddevice>=0.3.7,<0.4
click>=6.7,<7
tenacity>=4.1.0,<5
futures>=3.1.1,<4
pathlib2>=2.3.0,<3
google-assistant-library
google-assistant-sdk
google-auth-oauthlib[tool]
PyQtWebEngine